# openapi/zenart.v1.yaml
openapi: 3.1.0
info:
  title: ZenArt API
  version: 0.1.0-stage0-rev2
  description: Stage 0 Rev2 API contract. Docs/stage0_blueprint_rev2.md is authoritative.
servers:
  - url: /api/v1
    description: User API
  - url: /api/admin/v1
    description: Admin API
security:
  - ZenArtSession: []
...
paths:
...
  /tasks/{task_id}:
    get:
      operationId: getTask
      tags: [tasks]
      x-rbac: user
      parameters:
        - $ref: "#/components/parameters/TaskId"
      responses:
        "200":
          description: Long-running task status.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskStatus"
        default:
          $ref: "#/components/responses/Error"
...
  /packages/{package_id}/exports:
    post:
      operationId: createExport
      tags: [packages]
      x-rbac: user
      x-idempotency-required: true
      parameters:
        - $ref: "#/components/parameters/PackageId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExportCreate"
      responses:
        "202":
          description: Export task accepted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskStatus"
        default:
          $ref: "#/components/responses/Error"
  /exports/{export_id}:
    get:
      operationId: getExport
      tags: [packages]
      x-rbac: user
      parameters:
        - $ref: "#/components/parameters/ExportId"
      responses:
        "200":
          description: Export record.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Export"
        default:
          $ref: "#/components/responses/Error"
...
  /support/tickets:
    get:
      operationId: listSupportTickets
      tags: [support]
      x-rbac: admin
      parameters:
        - $ref: "#/components/parameters/PageToken"
        - $ref: "#/components/parameters/PageSize"
        - $ref: "#/components/parameters/Sort"
        - $ref: "#/components/parameters/Search"
        - $ref: "#/components/parameters/StatusFilter"
      responses:
        "200":
          description: Support ticket page for admin review and triage.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupportTicketPage"
        default:
          $ref: "#/components/responses/Error"
    post:
      operationId: createSupportTicket
      tags: [support]
      x-rbac: user
      x-idempotency-required: true
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SupportTicketCreate"
      responses:
        "201":
          description: Support ticket.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupportTicket"
        default:
          $ref: "#/components/responses/Error"
  /exports:
    get:
      operationId: listExports
      tags: [exports]
      x-rbac: admin
      parameters:
        - $ref: "#/components/parameters/PageToken"
        - $ref: "#/components/parameters/PageSize"
        - $ref: "#/components/parameters/Sort"
        - $ref: "#/components/parameters/Search"
        - $ref: "#/components/parameters/StatusFilter"
      responses:
        "200":
          description: Export job page for admin support and operations.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExportPage"
        default:
          $ref: "#/components/responses/Error"
  /exports/{export_id}/regenerate:
    post:
      operationId: regenerateExport
      tags: [exports]
      x-rbac: admin
      x-idempotency-required: true
      parameters:
        - $ref: "#/components/parameters/ExportId"
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "202":
          description: Export regeneration accepted for admin support.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Export"
        default:
          $ref: "#/components/responses/Error"
...
  /crawler/sources:
    get:
      operationId: listCrawlerSources
      tags: [crawler]
      x-rbac: admin
      parameters:
        - $ref: "#/components/parameters/PageToken"
        - $ref: "#/components/parameters/PageSize"
        - $ref: "#/components/parameters/StatusFilter"
        - $ref: "#/components/parameters/Search"
      responses:
        "200":
          description: Crawler source page.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CrawlerSourcePage"
        default:
          $ref: "#/components/responses/Error"
...
  /crawler/findings:
    get:
      operationId: listCrawlerFindings
      tags: [crawler]
      x-rbac: admin
      parameters:
        - $ref: "#/components/parameters/PageToken"
        - $ref: "#/components/parameters/PageSize"
        - $ref: "#/components/parameters/StatusFilter"
        - $ref: "#/components/parameters/Search"
      responses:
        "200":
          description: Crawler finding page.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CrawlerFindingPage"
        default:
          $ref: "#/components/responses/Error"
...
  /safety/rules:
    get:
      operationId: listSafetyRules
      tags: [safety]
      x-rbac: admin
      parameters:
        - $ref: "#/components/parameters/PageToken"
        - $ref: "#/components/parameters/PageSize"
        - $ref: "#/components/parameters/StatusFilter"
        - $ref: "#/components/parameters/Search"
      responses:
        "200":
          description: Safety rule page.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SafetyRulePage"
        default:
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    ZenArtSession:
      type: apiKey
      in: header
      name: X-Zenart-User-ID
      description: Stage 0 local session header. Production auth must replace this with signed session validation.
  parameters:
    PageToken:
      name: page_token
      in: query
      schema:
        type: string
      description: Cursor token returned by the previous page.
    PageSize:
      name: page_size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 25
      description: Maximum records to return.
    Sort:
      name: sort
      in: query
      schema:
        type: string
        example: "-created_at"
      description: Comma-separated sort keys. Prefix with - for descending.
    Search:
      name: search
      in: query
      schema:
        type: string
        maxLength: 200
      description: Search text applied to indexed text fields.
    StatusFilter:
      name: status
      in: query
      schema:
        type: string
      description: Filter by resource status.
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 8
        maxLength: 200
      description: Required for mutating requests so retries are safe.
...
    TaskId:
      name: task_id
      in: path
      required: true
      schema:
        type: string
...
    PackageId:
      name: package_id
      in: path
      required: true
      schema:
        type: string
    ExportId:
      name: export_id
      in: path
      required: true
      schema:
        type: string
...
  responses:
    Error:
      description: Shared error envelope.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
    ErrorEnvelope:
      type: object
      required: [code, message, request_id, details, field_errors]
      properties:
        code:
          type: string
        message:
          type: string
        request_id:
          type: string
        details:
          type: object
          additionalProperties: true
        field_errors:
          type: array
          items:
            $ref: "#/components/schemas/FieldError"
    FieldError:
      type: object
      required: [field, message]
      properties:
        field:
          type: string
        message:
          type: string
        code:
          type: string
    PageInfo:
      type: object
      required: [next_page_token, total_count]
      properties:
        next_page_token:
          type: string
        total_count:
          type: integer
          minimum: 0
...
    TaskStatus:
      type: object
      required:
        - id
        - tenant_id
        - type
        - schema_version
        - status
        - progress
        - retry_count
        - timeout_at
        - user_message
        - app_version
        - worker_version
        - created_at
        - updated_at
      properties:
        id:
          type: string
        tenant_id:
          type: string
        type:
          type: string
        schema_version:
          type: integer
        status:
          type: string
          enum: [pending, running, succeeded, failed, cancelled]
        progress:
          type: number
          minimum: 0
          maximum: 100
        retry_count:
          type: integer
          minimum: 0
        timeout_at:
          type: string
          format: date-time
          nullable: true
        error:
          $ref: "#/components/schemas/TaskError"
        user_message:
          type: string
        app_version:
          type: string
        worker_version:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    TaskError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
...
    ObjectMetadata:
      type: object
      required: [id, bucket, object_key, content_type, byte_size, checksum, created_at]
      properties:
        id:
          type: string
        bucket:
          type: string
        object_key:
          type: string
        content_type:
          type: string
        byte_size:
          type: integer
          minimum: 0
        checksum:
          type: string
        retention_until:
          type: string
          format: date-time
          nullable: true
        metadata:
          type: object
          additionalProperties: true
        created_at:
          type: string
          format: date-time
...
    ExportCreate:
      type: object
      required: [format]
      properties:
        format:
          type: string
          enum: [zip, pdf]
    Export:
      type: object
      required:
        - id
        - package_id
        - format
        - status
        - qa_status
        - manifest
        - qa_report
        - provenance
        - download_url
        - created_at
      properties:
        id:
          type: string
        tenant_id:
          type: string
        package_id:
          type: string
        task_id:
          type: string
          nullable: true
        format:
          type: string
        status:
          type: string
        qa_status:
          type: string
        object_metadata:
          $ref: "#/components/schemas/ObjectMetadata"
        manifest:
          type: object
          additionalProperties: true
        qa_report:
          type: object
          additionalProperties: true
        provenance:
          type: object
          additionalProperties: true
        download_url:
          type: string
          format: uri
          nullable: true
          description: Short-lived signed URL for export download when the object is ready.
        download_expires_at:
          type: string
          format: date-time
          nullable: true
        error:
          type: object
          additionalProperties: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ExportPage:
      allOf:
        - $ref: "#/components/schemas/PageInfo"
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/Export"
...
    SupportTicketCreate:
      type: object
      required: [category, body]
      properties:
        category:
          type: string
        body:
          type: string
        linked_export_id:
          type: string
    SupportTicket:
      type: object
      required: [id, tenant_id, user_id, category, status, body, created_at, updated_at]
      properties:
        id:
          type: string
        tenant_id:
          type: string
        user_id:
          type: string
        project_id:
          type: string
          nullable: true
        category:
          type: string
        status:
          type: string
        body:
          type: string
        linked_export_id:
          type: string
          nullable: true
        metadata:
          type: object
          additionalProperties: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    SupportTicketPage:
      allOf:
        - $ref: "#/components/schemas/PageInfo"
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/SupportTicket"
...
    CrawlerSource:
      type: object
      required: [id, name, url, approval_status, legal_metadata, robots_policy]
      properties:
        id:
          type: string
        name:
          type: string
        url:
          type: string
        approval_status:
          type: string
        legal_metadata:
          type: object
          additionalProperties: true
        robots_policy:
          type: object
          required: [robots_allowed, direct_activation_allowed]
          properties:
            robots_allowed:
              type: boolean
            direct_activation_allowed:
              type: boolean
              const: false
    CrawlerSourcePage:
      allOf:
        - $ref: "#/components/schemas/PageInfo"
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/CrawlerSource"
...
    CrawlerFinding:
      type: object
      required: [id, finding_type, status, payload, provenance, import_governance]
      properties:
        id:
          type: string
        finding_type:
          type: string
        status:
          type: string
        payload:
          type: object
          additionalProperties: true
        provenance:
          type: object
          additionalProperties: true
        import_governance:
          type: object
          required: [direct_activation_allowed, provenance_links_required, admin_review_required]
          properties:
            direct_activation_allowed:
              type: boolean
              const: false
            provenance_links_required:
              type: boolean
              const: true
            admin_review_required:
              type: boolean
              const: true
    CrawlerFindingPage:
      allOf:
        - $ref: "#/components/schemas/PageInfo"
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/CrawlerFinding"
...
    SafetyRule:
      type: object
      required: [id, rule_key, version, domain, severity, action, enforcement_points, evaluation_contract, status]
      properties:
        id:
          type: string
        rule_key:
          type: string
        version:
          type: string
        domain:
          type: string
        severity:
          type: string
        action:
          type: string
        enforcement_points:
          type: array
          items:
            type: string
            enum: [brief, provider_request, provider_response, qa, export]
          minItems: 5
          uniqueItems: true
        evaluation_contract:
          type: object
          required: [decision_schema, blocks_export_when_critical]
          properties:
            decision_schema:
              type: string
              const: SafetyDecision
            blocks_export_when_critical:
              type: boolean
              const: true
        status:
          type: string
    SafetyRulePage:
      allOf:
        - $ref: "#/components/schemas/PageInfo"
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/SafetyRule"
...